Polymorphize input_device and input_class; move to separate source file (nw)
